requirements.txt
boto3==1.17.29
botocore==1.20.29
jmespath==0.10.0
numpy==1.20.1
pandas==1.2.3
psycopg2-binary==2.8.6
python-dateutil==2.8.1
pytz==2021.1
s3transfer==0.3.4
six==1.15.0
urllib3==1.26.4
